minor styling, renaming and default enabled
