+++ examples/api-samples/src/browser/toolbar/sample-toolbar-contribution.css
@@ -14,30 +14,30 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-#theia-easy-search-toolbar-widget {
+#theia-sample-toolbar-contribution {
     position: relative;
 }
 
-#theia-easy-search-toolbar-widget .icon-wrapper {
+#theia-sample-toolbar-contribution .icon-wrapper {
     cursor: pointer;
     margin-left: 0;
 }
 
-#theia-easy-search-toolbar-widget:focus,
-#theia-easy-search-toolbar-widget .icon-wrapper:focus,
-#theia-easy-search-toolbar-widget .codicon-search:focus {
+#theia-sample-toolbar-contribution:focus,
+#theia-sample-toolbar-contribution .icon-wrapper:focus,
+#theia-sample-toolbar-contribution .codicon-search:focus {
     outline: none;
 }
 
-#theia-easy-search-toolbar-widget  .icon-wrapper.action-item.item.enabled:hover {
+#theia-sample-toolbar-contribution  .icon-wrapper.action-item.item.enabled:hover {
     background-color: var(--theia-toolbar-hoverBackground);
 }
 
-#theia-easy-search-toolbar-widget #easy-search-item-icon.codicon-search {
+#theia-sample-toolbar-contribution #easy-search-item-icon.codicon-search {
     position: relative;
 }
 
-#theia-easy-search-toolbar-widget .icon-wrapper .codicon-triangle-down {
+#theia-sample-toolbar-contribution .icon-wrapper .codicon-triangle-down {
     position: absolute;
     font-size: 10px;
     bottom: -7px;
